add change cart and CartServiseImpl and PriceDalIml

+++ src/main/java/service/CartServiceImpl.java
@@ -53,7 +53,7 @@ public class CartServiceImpl implements CartService {
     @Override
     @Transactional
     public Optional<Cart> createCart(Cart cart) {
-        Long cartId = cart.getId();
+        Integer cartId = cart.getId();
         boolean cartIdIsInDB = false;
         List<Cart> cartList = cartDal.readAllFromDB();
         for (Cart cart1 : cartList) {
